Chatbot1/requirements.txt
python-telegram-bot==21.10
APScheduler==3.6.3
google-generativeai==0.3.2
pymongo==4.6.1
python-dotenv==1.0.0
pillow>=10.0.0
tzdata==2024.1
